uncentralized getkeys.py
